Starting to refactor the code. Pulled logic out to SwiftypeMiddlemanHelper class and implemented it in the Thor subcommand. Put placeholder for the after_build event.
