+++ lib/middleman-swiftype/commands.rb
@@ -1,6 +1,6 @@
-require "middleman-core/cli"
-
-require "middleman-swiftype/pkg-info"
+require 'middleman-core/cli'
+require 'middleman-swiftype/pkg-info'
+require 'middleman-swiftype-helper'
 
 PACKAGE = "#{Middleman::Swiftype::PACKAGE}"
 VERSION = "#{Middleman::Swiftype::VERSION}"
@@ -32,7 +32,13 @@ module Middleman
         :desc => "Generate a search.json file without pushing it"
 
       def swiftype
-        if options[:"only-generate"]
+        shared_instance = Middleman::Application.server.inst
+        generate_only = options[:"only-generate"]
+        plugin_options = swiftype_options(shared_instance, generate_only)
+        helper = MiddlemanSwiftypeHelper.new plugin_options, shared_instance
+
+        if generate_only
+          #TODO: remove this once after_build handler is complete
           shared_instance.logger.info("Building content...")
           builder = Middleman::Cli::Build.new
           builder.build
@@ -40,11 +46,11 @@ module Middleman
           shared_instance.logger.info("Done. Creating search.json...")
           File.open("./#{Middleman::Application.build_dir}/search.json", "w") do |f|
             f.write("{\"documents\": ")
-            f.write(self.generate_swiftype_records.to_json)
+            f.write(helper.generate_swiftype_records.to_json)
             f.write("}")
           end
         else
-          self.push_to_swiftype(self.generate_swiftype_records)
+          helper.push_to_swiftype(helper.generate_swiftype_records)
         end
       end
 
@@ -71,11 +77,7 @@ end
 EOF
       end
 
-      def swiftype_options(shared_instance, generate_only=false)
-        require 'swiftype'
-        require 'nokogiri'
-        require 'digest'
-
+      def swiftype_options(shared_instance, generate_only)
         options = nil
 
         begin
@@ -95,117 +97,6 @@ EOF
         end
 
         options
-      end
-
-      def generate_swiftype_records
-        records = []
-
-        options = self.swiftype_options(shared_instance, true)
-        m_pages = shared_instance.sitemap.resources.find_all{|p| options.pages_selector.call(p) }
-
-        m_pages.each do |p|
-          external_id = Digest::MD5.hexdigest(p.url)
-
-          # optional selector for retrieving the page title
-          if options.title_selector
-            title = options.title_selector.call(shared_instance, p)
-          else
-            title = p.metadata[:page]['title']
-          end
-
-          url = p.url
-          sections = []
-          body = ''
-          info = ''
-          image = ''
-
-          f = Nokogiri::HTML.fragment(p.render(:layout => false))
-
-          # optionally edit html
-          if options.process_html
-            options.process_html.call(f)
-          end
-          body = f.text
-
-          if options.generate_sections
-            sections = options.generate_sections.call(p)
-          end
-
-          # optionally generate extra info
-          if options.generate_info
-            info = options.generate_info.call(f)
-          end
-
-          # optional image
-          if options.generate_image
-            image = options.generate_image.call(p)
-          end
-
-          if options.should_index
-            should_index = options.should_index.call(p, title)
-            next unless should_index
-          end
-
-          fields = [
-            {:name => 'title', :value => title, :type => 'string'},
-            {:name => 'url', :value => url, :type => 'enum'},
-            {:name => 'body', :value => body, :type => 'string'},
-            {:name => 'info', :value => info, :type => 'string'}
-          ]
-
-          if sections.length > 0
-            {:name => 'sections', :value => sections, :type => 'string'}
-          end
-
-          if image
-            fields << {:name => 'image', :value => image, :type => 'enum'}
-          end
-
-          records << {
-            :external_id => external_id,
-            :fields => fields
-          }
-        end
-
-        records
-      end
-
-      def push_to_swiftype(records)
-        options = self.swiftype_options(shared_instance)
-        # https://github.com/swiftype/swiftype-rb
-        ::Swiftype.configure do |config|
-          config.api_key = options.api_key
-        end
-
-        swiftype_client = ::Swiftype::Client.new
-
-        records.each do |record|
-          # https://swiftype.com/documentation/crawler#schema
-          # https://swiftype.com/documentation/meta_tags
-          url_field = record[:fields].find { |fields| fields[:name] == "url" }
-          shared_instance.logger.info("Pushing contents of #{url_field[:value]} to swiftype")
-          #next
-          begin
-            swiftype_client.create_or_update_document(options.engine_slug, swiftype_document_type, {
-                :external_id => record[:external_id],
-                :fields => record[:fields]
-            })
-          rescue ::Swiftype::NonExistentRecord
-            swiftype_client.create_document_type(options.engine_slug, swiftype_document_type)
-            swiftype_client.create_or_update_document(options.engine_slug, 'page', {
-                :external_id => record[:external_id],
-                :fields => record[:fields]
-            })
-          end
-        end
-      end
-
-      def shared_instance
-        @shared_instance ||= Middleman::Application.server.inst
-      end
-
-      def swiftype_document_type
-        "page"
       end
     end
   end
